포트원 결제 완료

+++ src/main/resources/db/table.sql
@@ -126,7 +126,7 @@ create table admin_prod_history (
     end_time int (2) ,
     cancel_yn varchar (1) default 'N',
     cancel_at timestamp ,
-    cancel_amount int (7) ,
+    cancel_amount double,
     people_count int (3),
     purchase_date varchar (100)
 );
